Added required static files for the editor.
The editor now includes a CSS file and some icon files. All files and
directories in the build directory should now be included in production.
The editor should figure out it's relative URL and include the CSS file
as appropriately from that.

Currently any edits or additions to CSS and image files in the source
directory needs to be manually copied to the build directory when
appropriate. In the future the build script will do this automatically.